Added new app-tier service and pod
A new Kubernetes Service and Pod have been added for the application tier. The service exposes port 8080 and is labeled as part of the 'microservices' app. The corresponding Pod runs a server container from a specific image, also exposing port 8080. Environment variables are set for Redis URL using data tier service host and port. Resource requests and limits are defined for memory and CPU usage.