Added metrics-server component to Kubernetes
This commit introduces the metrics-server component to the Kubernetes system. The new file 'components.yaml' includes definitions for a ServiceAccount, ClusterRoles, RoleBinding, ClusterRoleBindings, a Service and a Deployment related to the metrics-server. It also sets up an APIService for v1beta1.metrics.k8s.io. This will allow for resource usage data collection from each node and pod in the cluster.